Dodate metode za racunanje prosecne ocene klinike i doktora

+++ src/main/java/com/ftn/isa/dto/request/ClinicReviewRequest.java
@@ -13,7 +13,7 @@ public class ClinicReviewRequest {
 
     private Long patientId;
 
-    private int review;
+    private Double review;
 
     private String description;
 }
