+++ src/main/java/com/ftn/isa/dto/request/DoctorReviewRequest.java
@@ -13,7 +13,7 @@ public class DoctorReviewRequest {
 
     private Long patientId;
 
-    private int review;
+    private Double review;
 
     private String description;
 }
