+++ src/main/java/com/ftn/isa/dto/response/ClinicReviewResponse.java
@@ -15,7 +15,7 @@ public class ClinicReviewResponse {
 
     private Long patientId;
 
-    private int review;
+    private Double review;
 
     private String description;
