+++ src/main/java/com/ftn/isa/dto/response/DoctorReviewResponse.java
@@ -15,7 +15,7 @@ public class DoctorReviewResponse {
 
     private Long patientId;
 
-    private int review;
+    private Double review;
 
     private String description;
 }
